# docs/api/openapi-legalpro-node.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LegalPro Node API",
    "version": "1.0.0",
    "description": "API backend Node.js de LegalPro - auth, organizaciones, IA, minimax, ARCO, datos personales",
    "contact": { "name": "LegalPro", "email": "api@legalpro.pe" }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Dev local" },
    { "url": "https://legalpro-node.railway.app", "description": "Produccion" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "nombre_completo": { "type": "string" },
          "rol": { "enum": ["ABOGADO", "FISCAL", "JUEZ", "CONTADOR", "ADMIN"] },
          "acepta_transferencia_internacional": { "type": "boolean" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "accessToken": { "type": "string" },
              "refreshToken": { "type": "string" },
              "user": { "$ref": "#/components/schemas/Usuario" }
            }
          }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Login con email y password",
        "tags": ["auth"],
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } },
          "401": { "description": "Credenciales invalidas" },
          "429": { "description": "Demasiados intentos (brute force)" }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Refrescar access token",
        "tags": ["auth"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "refreshToken": { "type": "string" } } } } } },
        "responses": { "200": { "description": "OK" }, "401": { "description": "Refresh invalido" } }
      }
    },
...
    "/api/expedientes/{id}": {
      "get": { "summary": "Obtener expediente por ID", "tags": ["expedientes"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "OK" }, "404": { "description": "No encontrado" } } }
    },
...
  }
}